splitting store
