+++ src/store/modules/accounts-pending-txs/accounts-pending-txs-actions.js
@@ -1,0 +1,3 @@
+export default {
+
+}
