+++ src/store/modules/accounts-pending-txs/accounts-pending-txs-store.js
@@ -1,0 +1,15 @@
+import mutations from "./accounts-pending-txs-mutations"
+import actions from "./accounts-pending-txs-actions"
+
+export default {
+
+    state: {
+
+        list: {},
+
+    },
+
+    actions,
+    mutations,
+
+}
