+++ src/store/modules/accounts-txs/accounts-txs-store.js
@@ -1,0 +1,16 @@
+import mutations from "./accounts-txs-mutations"
+import actions from "./accounts-txs-actions"
+
+export default {
+
+    state: {
+
+        list: {},
+        viewTxsPositions: {},
+
+    },
+
+    actions,
+    mutations,
+
+}
