+++ src/store/store.js
@@ -3,6 +3,8 @@ import Vuex from 'vuex';
 import WalletStore from "./modules/wallet/wallet-store"
 import AddressesStore from "./modules/addresses/addresses-store"
 import AccountsStore from "./modules/accounts/accounts-store"
+import AccountsTxsStore from "./modules/accounts-txs/accounts-txs-store"
+import AccountsPendingTxsStore from "./modules/accounts-pending-txs/accounts-pending-txs-store"
 import BlockchainStore from "./modules/blockchain/blockchain-store"
 import BlocksStore from "./modules/blocks/blocks-store"
 import BlocksInfoStore from "./modules/blocks-info/blocks-info-store"
@@ -33,6 +35,8 @@ const store = new Vuex.Store({
         tokensInfo: TokensInfoStore,
         addresses: AddressesStore,
         accounts: AccountsStore,
+        accountsTxs: AccountsTxsStore,
+        accountsPendingTxs: AccountsPendingTxsStore,
         mempool: MemPoolStore,
         faucet: FaucetStore,
         toasts: ToastsStore,
